+++ packages/cm6-graphql/CHANGELOG.md
@@ -1,5 +1,12 @@
 # cm6-graphql
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [[`4d33b221`](https://github.com/graphql/graphiql/commit/4d33b2214e941f171385a1b72a1fa995714bb284)]:
+  - graphql-language-service@5.1.5
+
 ## 0.0.5
 
 ### Patch Changes