+++ packages/graphql-language-service-cli/CHANGELOG.md
@@ -1,5 +1,20 @@
 # graphql-language-service-cli
 
+## 3.3.21
+
+### Patch Changes
+
+- Updated dependencies [[`f2040452`](https://github.com/graphql/graphiql/commit/f20404529677635f5d4792b328aa648641bf8d9c)]:
+  - graphql-language-service-server@2.10.0
+
+## 3.3.20
+
+### Patch Changes
+
+- Updated dependencies [[`4d33b221`](https://github.com/graphql/graphiql/commit/4d33b2214e941f171385a1b72a1fa995714bb284)]:
+  - graphql-language-service-server@2.9.10
+  - graphql-language-service@5.1.5
+
 ## 3.3.19
 
 ### Patch Changes